+++ tests/test.js
@@ -69,6 +69,14 @@ test('viewing the details of a rental property', async ({ page }) => {
 
 	// Click on a rental
 	await page.locator('text=Grand Old Mansion').click();
+
+	// Assert that the share button is visible and has the correct attributes
+	expect(await page.textContent('.share.button')).toBe('Share on Twitter');
+	const href = await page.getAttribute('.share.button', 'href');
+	const tweetURL = new URL(href);
+	expect(tweetURL.host).toBe('twitter.com');
+	expect(tweetURL.searchParams.get('url')).toBe(`${[page.url()]}`);
+
 	// Assert it has taken you to the rental page
 	await expect(page).toHaveURL('/rentals/grand-old-mansion');
 });